+++ src/main/java/com/buuz135/industrial/proxy/client/ItemRenderRegistry.java
@@ -20,5 +20,6 @@ public class ItemRenderRegistry {
 
         adultFilterAddomItem.registerRenderer();
         rangeAddonItem.registerRenderer();
+        leafShearingAddonItem.registerRenderer();
     }
 }